requirements.txt
Django==4.2.10
djangorestframework==3.14.0
pre-commit==3.5.0
drf-spectacular==0.27.1
pytest-django==4.8.0
factory_boy==3.3.0
